dynamic programming working but not efficient

+++ src/MainFunctions/OtherHeuristics.h
@@ -5,8 +5,8 @@
 
 using namespace std;
     vector<int> cluster_graph(DataManager aux, int k);
-    pair<int,vector<pair<int,Vertex *>>> tsp_dynamic_programming(vector<vector<double>> distances, vector<pair<int,Vertex *>> vertices, int pos, int mask, int k, pair<int,vector<pair<int, Vertex *>>> path);
-    pair<int,vector<pair<int, Vertex *>>> joinSets(pair<int,vector<pair<int, Vertex *>>> set1, pair<int,vector<pair<int, Vertex *>>> set2, bool toy);
-    int OtherHeuristic(DataManager aux, int k, bool toy);
-    vector<vector<double>> getDistances(vector<pair<int, Vertex*>> vertices, bool toy);
+    pair<double,vector<pair<int,Vertex *>>> tsp_dynamic_programming(vector<vector<double>> distances, vector<pair<int,Vertex *>> vertices, int pos, int mask, int k, pair<int,vector<pair<int, Vertex *>>> path);
+    pair<double,vector<pair<int, Vertex *>>> joinSets(pair<double,vector<pair<int, Vertex *>>> set1, pair<double,vector<pair<int, Vertex *>>> set2, bool toy);
+    double OtherHeuristic(DataManager aux, int k, bool toy);
+    vector<vector<double>> getDistances(vector<pair<double, Vertex*>> vertices, bool toy);
 #endif
